reduced file count
