--- malformer/utils/defaults.py
@@ -1,6 +1,0 @@
-ACCOUNT = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NDg1ODI0NjUsIm5iZiI6MTcxNzA0NjQ2NSwiaWF0IjoxNzE3MDQ2NDY1LCJpZGVudGl0eSI6IjBkZWNkZGNmZGYyNDFhNjAifQ.g3khyJgOkNvZny6Vh579cuQj1HLLGSDeauZbfZri9jw'
-DEVICE = '1d3dc3e03047b0c7'
-ROUTE = '000000dd--455f14369d'
-FPS = 20 # video frame rate, for performing calculations on qlogs
-STORAGE_PATH = './malformer/routes'
-CAPNP_MAP = { 'GPS': 'gpsLocation', 'Thumbnail': 'thumbnail', 'qlogs': 'qEncodeIdx' }
